Return firefox_path if firefox_path is not None

+++ tests/ui/conftest.py
@@ -63,30 +63,29 @@ def user(base_url, fxa_account, jwt_token):
 
 
 @pytest.yield_fixture(scope='session')
-def firefox_path(tmpdir_factory):
-    tmp_dir = tmpdir_factory.mktemp('firefox')
-    scraper = FactoryScraper('release', version='latest')
-    filename = scraper.download()
-    path = mozinstall.install(filename, str(tmp_dir))
+def firefox_path(tmpdir_factory, firefox_path):
+    if firefox_path is not None:
+        yield firefox_path
+    else:
+        tmp_dir = tmpdir_factory.mktemp('firefox')
+        scraper = FactoryScraper('release', version='latest', destination='tmp_dir')
+        filename = scraper.download()
+        path = mozinstall.install(filename, str(tmp_dir))
 
-    yield mozinstall.get_binary(path, 'Firefox')
+        yield mozinstall.get_binary(path, 'Firefox')
 
-    mozinstall.uninstall(path)
-    os.rmdir(str(tmp_dir))
-    os.remove(filename)
+        mozinstall.uninstall(path)
+        os.remove(filename)
+        os.rmdir(str(tmp_dir))
 
 
 @pytest.fixture
 def discovery_pane_url(base_url):
-    disover_url = ''
-
     if 'localhost' in base_url:
         discover_url = None
     elif 'dev' in base_url:
-        discover_url = 'https://discovery.addons-dev.allizom.org/'
+        return 'https://discovery.addons-dev.allizom.org/'
     elif 'allizom' in base_url:
-        discover_url = 'https://discovery.addons.allizom.org/'
+        return 'https://discovery.addons.allizom.org/'
     else:
-        discover_url = 'https://discovery.addons.mozilla.org/'
-
-    return discover_url
+        return 'https://discovery.addons.mozilla.org/'
